Factor out lib/serialization.ts.
