+++ lib/pages/mandala-page/serialization.test.ts
@@ -1,19 +1,8 @@
 import {
-  ColorPacker,
   serializeMandalaDesign,
   deserializeMandalaDesign,
 } from "./serialization";
 import { MANDALA_DESIGN_DEFAULTS } from "./core";
-
-describe("AvroColorConverter", () => {
-  it("converts strings to numbers", () => {
-    expect(ColorPacker.pack("#abcdef")).toEqual(0xabcdef);
-  });
-
-  it("converts numbers to strings", () => {
-    expect(ColorPacker.unpack(0xabcdef)).toEqual("#abcdef");
-  });
-});
 
 describe("Mandala design serialization/desrialization", () => {
   // Helper to make it easy for us to copy/paste from URLs.
